pyproject.toml
[tool.poetry]
name = "aadg-genomics-class"
version = "0.1.0"
description = "Implementation of read mapping algorithm for the genomics classes @ MIMUW 2023"
authors = ["styczynski <piotr@styczynski.in>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
tqdm = "^4.66.1"
numpy = "^1.26.2"
biopython = "^1.81"
pydantic = "^2.5.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
